safari-600.7-branch/LayoutTests/css3/font-feature-settings-parsing-expected.txt

Test parsing of the CSS3 font-feature-settings property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


- Tests valid inputs.
PASS parseResultOf("valid_normal") is "normal"
PASS parseResultOf("valid_value_1") is "'dlig' 1"
PASS parseResultOf("valid_value_2") is "'swsh' 2"
PASS parseResultOf("valid_value_on") is "'smcp' 1"
PASS parseResultOf("valid_value_off") is "'liga' 0"
PASS parseResultOf("valid_value_omit") is "'c2sc' 1"
PASS parseResultOf("valid_valuelist") is "'tnum' 1, 'hist' 1"
PASS parseResultOf("valid_singlequote") is "'PKRN' 1"
PASS parseResultOf("valid_unusual_tag") is "'!@#$' 1"
PASS parseResultOf("valid_tag_space") is "'a bc' 1"
PASS parseResultOf("valid_composite") is "'dlig' 1, 'smcp' 1, 'lig ' 0"
- Tests invalid inputs.  Results should be "normal".
PASS parseResultOf("invalid_ident") is "normal"
PASS parseResultOf("invalid_cases") is "normal"
PASS parseResultOf("invalid_1") is "normal"
PASS parseResultOf("invalid_off") is "normal"
PASS parseResultOf("invalid_normal_duplicate") is "normal"
PASS parseResultOf("invalid_normal_list") is "normal"
PASS parseResultOf("invalid_longer_ident") is "normal"
PASS parseResultOf("invalid_longer_quote") is "normal"
PASS parseResultOf("invalid_negative") is "normal"
PASS parseResultOf("invalid_float") is "normal"
PASS parseResultOf("invalid_missing_comma") is "normal"
PASS parseResultOf("invalid_missing_comma_ident") is "normal"
PASS parseResultOf("invalid_comma") is "normal"
PASS parseResultOf("invalid_ending_comma") is "normal"
PASS parseResultOf("invalid_beginning_comma") is "normal"
PASS parseResultOf("invalid_on") is "normal"
PASS parseResultOf("invalid_0") is "normal"
- Tests inherit.
PASS parseResultOf("outer") is "'dlig' 1"
PASS parseResultOf("inner") is "'dlig' 1"
- Tests @font-face.
PASS fontFaceRuleValid is "'liga' 1"
PASS fontFaceRuleInvalid is ""
PASS successfullyParsed is true

TEST COMPLETE
